+++ app/src/main/java/woowacourse/shopping/di/annotation/Inject.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.di.injector
+package woowacourse.shopping.di.annotation
 
 @Target(AnnotationTarget.PROPERTY)
 annotation class Inject
