+++ app/src/main/java/woowacourse/shopping/di/annotation/Qualifier.kt
@@ -1,0 +1,4 @@
+package woowacourse.shopping.di.annotation
+
+@Target(AnnotationTarget.PROPERTY)
+annotation class Qualifier(val implementation: String)
